change modifying variable default to private create interface which is responsible for CRUD (https://ru.wikipedia.org/wiki/CRUD) spring.xml create configuration file, create registration.jsp, registration.css create tests JUnit for class UserDAOImpl change configuration file, create configuration file for tests create interface Service for business logic to parameter change configuration file, tests, and create separate DAO for User and Role and their implementation change EntityManager to SessionFactory return EntityManager

+++ src/test/java/com/vardanian/service/impl/RoleDAOImplTest.java
@@ -36,11 +36,6 @@ public class RoleDAOImplTest {
 
     }
 
-    @After
-    public void tearDown() throws Exception {
-       roleService.remove(role);
-    }
-
     @Test
     public void testCreate() {
         roleService.create(role);
@@ -75,8 +70,11 @@ public class RoleDAOImplTest {
     public void testRemove() {
         Role checkRole = roleService.findByName("admin");
         assertNotNull(checkRole);
-        roleService.remove(checkRole);
+        if (checkRole != null) {
+            roleService.remove(checkRole);
+        }
         checkRole = roleService.findByName("admin");
+        assertEquals(null, checkRole);
     }
 
     @Test
@@ -89,7 +87,7 @@ public class RoleDAOImplTest {
     @Test
     public void testFindById() {
         roleService.create(role);
-        Role checkRole = roleService.findById(1L);
+        Role checkRole = roleService.findById(role.getId());
         assertEquals("1", checkRole.getId().toString());
     }
 }
